requirements.txt
Django
python-decouple==3.8   #for setting related security, protect sensitive data
pytz==2023.3   # for timezone aware datetime fields
dj-database-url==2.0.0   #for database url
psycopg2-binary==2.9.6   #for postgresql database
asgiref==3.7.2
django-cors-headers==4.2.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.8.0
pytz==2023.3
sqlparse
tzdata==2023.3
dnspython==2.4.1
django-filter==23.2
whitenoise==6.2.0
Brotli==1.0.9
python-dotenv==1.0.0
gunicorn==20.1.0
pymongo==3.12.3
djongo==1.3.6
pyairtable==2.2.1
srf-spectacular==0.26.5
